+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Align duration display in cp details with the one in cp list (#393)
 -   Increase toast font size (#395)
 -   Upgrade chartjs to 3.7.1 (#399)
+-   Make full PerfBrowser sidebar section clickable (#401)
 
 ## [0.12.0] - 2022-02-10